backend/requirements.txt
pandas
google-cloud-aiplatform
google-cloud-storage
google-adk
python-dotenv
vertexai
langchain
openpyxl.